Set up locator services to fix NPE during Maven priming
