+++ enterprise/micronaut/test/unit/src/org/netbeans/modules/micronaut/maven/MicronautPackagingArtifactImplTest.java
@@ -21,66 +21,34 @@ package org.netbeans.modules.micronaut.maven;
 import java.io.File;
 import java.util.ArrayList;
 import java.util.Collection;
-import java.util.Collections;
 import java.util.List;
-import java.util.Set;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
 import java.util.logging.Logger;
-import javax.swing.text.Document;
 import static junit.framework.TestCase.assertEquals;
 import static junit.framework.TestCase.assertNotNull;
 import org.netbeans.api.project.Project;
 import org.netbeans.api.project.ProjectActionContext;
 import org.netbeans.api.project.ProjectManager;
 import org.netbeans.api.project.ui.OpenProjects;
-import org.netbeans.junit.NbTestCase;
 import org.netbeans.modules.maven.api.MavenConfiguration;
-import org.netbeans.modules.parsing.impl.indexing.implspi.ActiveDocumentProvider;
+import org.netbeans.modules.micronaut.NbSuiteTestBase;
 import org.netbeans.modules.project.dependency.ArtifactSpec;
 import org.netbeans.modules.project.dependency.ProjectArtifactsQuery;
 import org.netbeans.spi.project.ActionProgress;
 import org.netbeans.spi.project.ActionProvider;
-import org.netbeans.spi.project.ProjectConfiguration;
 import org.netbeans.spi.project.ProjectConfigurationProvider;
 import org.openide.filesystems.FileObject;
 import org.openide.filesystems.FileUtil;
 import org.openide.modules.DummyInstalledFileLocator;
 import org.openide.util.lookup.Lookups;
-import org.openide.util.lookup.ServiceProvider;
 import org.openide.windows.IOProvider;
 
 /**
  *
  * @author sdedic
  */
-public class MicronautPackagingArtifactImplTest extends NbTestCase {
-    
-    // must register ADP: otherwise maven fails at the start and will not even run 
-    // Prime command.
-    @ServiceProvider(service = ActiveDocumentProvider.class)
-    public static class ActiveDocumentProviderImpl implements ActiveDocumentProvider {
-
-        @Override
-        public Document getActiveDocument() {
-            return null;
-        }
-
-        @Override
-        public Set<? extends Document> getActiveDocuments() {
-            return Collections.emptySet();
-        }
-
-        @Override
-        public void addActiveDocumentListener(ActiveDocumentListener listener) {
-        }
-
-        @Override
-        public void removeActiveDocumentListener(ActiveDocumentListener listener) {
-        }
-        
-    }
-
+public class MicronautPackagingArtifactImplTest extends NbSuiteTestBase {
     public MicronautPackagingArtifactImplTest(String name) {
         super(name);
     }
